fix: continue redis channel after 'no player found' error. (#356)
### Problem:

when there is no any user while admin starts quiz the error no player found is shown to admin and quiz doesn't start and admin continues wait for user to join.
after this error if any user join quiz the name of these will not shown to waiting area at admin side.

### solution implemented

the problem is because of redis channel is closed for listing users for admin side.
so the redis channel listen even after the 'no player found error' and stops when admin will start quiz with participants.

Reviewed-on: https://git.pride.improwised.dev/Improwised/quizz-app/pulls/356
Reviewed-by: ashvin.bambhaniya <ashvin.bambhaniya@improwised.com>
Co-authored-by: husen kureshi <husen.kureshi@improwised.com>
Co-committed-by: husen kureshi <husen.kureshi@improwised.com>